Extract logic to send visits to Matomo to its own service
