+++ module/Core/src/EventDispatcher/Matomo/SendVisitToMatomo.php
@@ -8,8 +8,7 @@ use Doctrine\ORM\EntityManagerInterface;
 use Psr\Log\LoggerInterface;
 use Shlinkio\Shlink\Core\EventDispatcher\Event\VisitLocated;
 use Shlinkio\Shlink\Core\Matomo\MatomoOptions;
-use Shlinkio\Shlink\Core\Matomo\MatomoTrackerBuilderInterface;
-use Shlinkio\Shlink\Core\ShortUrl\Helper\ShortUrlStringifier;
+use Shlinkio\Shlink\Core\Matomo\MatomoVisitSenderInterface;
 use Shlinkio\Shlink\Core\Visit\Entity\Visit;
 use Throwable;
 
@@ -18,9 +17,8 @@ readonly class SendVisitToMatomo
     public function __construct(
         private EntityManagerInterface $em,
         private LoggerInterface $logger,
-        private ShortUrlStringifier $shortUrlStringifier,
         private MatomoOptions $matomoOptions,
-        private MatomoTrackerBuilderInterface $trackerBuilder,
+        private MatomoVisitSenderInterface $visitSender,
     ) {
     }
 
@@ -42,48 +40,10 @@ readonly class SendVisitToMatomo
         }
 
         try {
-            $tracker = $this->trackerBuilder->buildMatomoTracker();
-
-            $tracker
-                ->setUrl($this->resolveUrlToTrack($visit))
-                ->setCustomTrackingParameter('type', $visit->type->value)
-                ->setUserAgent($visit->userAgent)
-                ->setUrlReferrer($visit->referer);
-
-            $location = $visit->getVisitLocation();
-            if ($location !== null) {
-                $tracker
-                    ->setCity($location->cityName)
-                    ->setCountry($location->countryName)
-                    ->setLatitude($location->latitude)
-                    ->setLongitude($location->longitude);
-            }
-
-            // Set not obfuscated IP if possible, as matomo handles obfuscation itself
-            $ip = $visitLocated->originalIpAddress ?? $visit->remoteAddr;
-            if ($ip !== null) {
-                $tracker->setIp($ip);
-            }
-
-            if ($visit->isOrphan()) {
-                $tracker->setCustomTrackingParameter('orphan', 'true');
-            }
-
-            // Send the short URL title or an empty document title to avoid different actions to be created by matomo
-            $tracker->doTrackPageView($visit->shortUrl?->title() ?? '');
+            $this->visitSender->sendVisitToMatomo($visit, $visitLocated->originalIpAddress);
         } catch (Throwable $e) {
             // Capture all exceptions to make sure this does not interfere with the regular execution
             $this->logger->error('An error occurred while trying to send visit to Matomo. {e}', ['e' => $e]);
         }
-    }
-
-    public function resolveUrlToTrack(Visit $visit): string
-    {
-        $shortUrl = $visit->shortUrl;
-        if ($shortUrl === null) {
-            return $visit->visitedUrl ?? '';
-        }
-
-        return $this->shortUrlStringifier->stringify($shortUrl);
     }
 }
