+++ module/Core/src/Util/RedirectStatus.php
@@ -16,7 +16,7 @@ enum RedirectStatus: int
         return contains($this, [self::STATUS_301, self::STATUS_308]);
     }
 
-    public function isLegacyStatus(): bool
+    public function isGetOnlyStatus(): bool
     {
         return contains($this, [self::STATUS_301, self::STATUS_302]);
     }
